基于XXL-JOB二次开发，更好的支持Spring Cloud架构

+++ xxl-job-common/src/main/java/com/xxl/job/model/InstanceRegistry.java
@@ -42,7 +42,7 @@ public class InstanceRegistry implements Serializable {
     /**
      * 权重
      */
-    private int weight = 1;
+    private Double weight = 1.0;
 
 
     /**
